requirements.txt
python-dotenv
playwright
google-generativeai
pytest
pytest-asyncio
SQLAlchemy
beautifulsoup4
requests
pandas
pymongo